requirements.txt
kiteconnect
pandas
psycopg2-binary
python-dotenv
sqlalchemy
schedule
